- Validation of JSX maximum depth - Non-interactive elements should not be assigned interactive roles
